demo/requirements.txt
Django==1.7.1
django-ajax-selects==1.3.5
